+++ src/devices/ZoneValve.ts
@@ -119,7 +119,7 @@ export class ZoneValve extends DeviceBase {
     })
 
     // Start an update interval
-    interval(this.platform.config.options!.refreshRate! * 1000)
+    interval(this.deviceRefreshRate * 1000)
       .pipe(skipWhile(() => this.zoneUpdateInProgress))
       .subscribe(() => {
         this.rainbird!.refreshStatus()
@@ -130,7 +130,7 @@ export class ZoneValve extends DeviceBase {
         tap(() => {
           this.zoneUpdateInProgress = true
         }),
-        debounceTime(this.platform.config.options!.pushRate! * 1000),
+        debounceTime(this.devicePushRate * 1000),
       )
       .subscribe(async (zone: number) => {
         try {